Move ChatAgentLocation to avoid import cycle. Update imports
