+++ src/vs/workbench/contrib/chat/common/chatParticipantContribTypes.ts
@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { RawChatParticipantLocation } from './constants.js';
+
 export interface IRawChatCommandContribution {
 	name: string;
 	description: string;
@@ -12,8 +14,6 @@ export interface IRawChatCommandContribution {
 	defaultImplicitVariables?: string[];
 	disambiguation?: { category: string; categoryName?: string /** Deprecated */; description: string; examples: string[] }[];
 }
-
-export type RawChatParticipantLocation = 'panel' | 'terminal' | 'notebook' | 'editing-session';
 
 export interface IRawChatParticipantContribution {
 	id: string;
